fourth trial

+++ NaverShoppingList_2ndRecap/ShoppingSearch/ShoppingListCollectionViewCell.swift
@@ -13,6 +13,8 @@ import Kingfisher
 
 class ShoppingListCollectionViewCell: UICollectionViewCell {
     
+    let numberFormatter: NumberFormatter = NumberFormatter()
+    
     var completionHandler: (() -> ())?
     let repository = ShoppingItemTableRepository()
     
@@ -33,7 +35,7 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
     
     let mallNameLabel = {
         let view = UILabel()
-        view.font = .systemFont(ofSize: 13)
+        view.font = .systemFont(ofSize: 18)
         view.textAlignment = .left
         view.textColor = .lightGray
         return view
@@ -41,7 +43,7 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
     
     let titleLabel = {
         let view = UILabel()
-        view.font = .systemFont(ofSize: 14)
+        view.font = .systemFont(ofSize: 18)
         view.textAlignment = .left
         view.textColor = .black
         view.numberOfLines = 2
@@ -50,7 +52,7 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
     
     let lpriceLabel = {
         let view = UILabel()
-        view.font = .systemFont(ofSize: 15, weight: .bold)
+        view.font = .systemFont(ofSize: 18, weight: .bold)
         view.textAlignment = .left
         view.textColor = .black
         return view
@@ -62,46 +64,57 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
     
     func showSearchedShoppingCellContents(repoShoppingList: Results<ShoppingItem>, searchedShoppingItem: Item, imageURL: URL) {
 
-        //MARK: - 버튼 이미지 삼항연산자로 빼보자
         if repoShoppingList.contains(where: { $0.productID == searchedShoppingItem.productID }) {
-            likeButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
+            likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartSelected),
+                                for: .normal)
         } else {
-            likeButton.setImage(UIImage(systemName: "heart"), for: .normal)
+            likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartNone),
+                                for: .normal)
         }
 
         let trimedTitle = searchedShoppingItem.title.components(separatedBy: ["<", "b", ">", "/"]).joined()
 
-
         mallNameLabel.text = searchedShoppingItem.mallName
         titleLabel.text = trimedTitle
-        lpriceLabel.text = searchedShoppingItem.lprice
+        
+        numberFormatter.numberStyle = .decimal
+        
+        guard let lPriceInt = Int(searchedShoppingItem.lprice), let result: String = numberFormatter.string(for: lPriceInt) else { return }
+                
+        lpriceLabel.text = result + "원"
         productImage.kf.setImage(with: imageURL)
 
         completionHandler = { [weak self] in
-            //MARK: - 여기서 searchedShoppingItem의 productID값
+            
             if repoShoppingList.contains(where: { $0.productID == searchedShoppingItem.productID }) {
 
                 self?.repository.deleteShoppingItem(savedItemList: repoShoppingList, searchedItemList: searchedShoppingItem)
 
-                self?.likeButton.setImage(UIImage(systemName: "heart"), for: .normal)
+                self?.likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartNone), for: .normal)
             } else {
                 self?.repository.createShoppingItem(itemToSave: searchedShoppingItem)
 
-                self?.likeButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
+                self?.likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartSelected), for: .normal)
             }
         }
     }
     
     func showLikeListCellContents(imageURL: URL, savedItem: Results<ShoppingItem>, index: Int) {
         productImage.kf.setImage(with: imageURL)
-        likeButton.setImage(UIImage(systemName: "heart.fill"), for: .normal)
-        likeButton.setImage(UIImage(systemName: "heart"), for: .highlighted)
+        likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartSelected), for: .normal)
+        likeButton.setImage(UIImage(systemName: ButtonImageName.shared.heartNone), for: .highlighted)
         mallNameLabel.text = savedItem[index].mallName
         titleLabel.text = savedItem[index].title
-        lpriceLabel.text = savedItem[index].lprice
+ 
+        numberFormatter.numberStyle = .decimal
+        
+        guard let lPriceInt = Int(savedItem[index].lprice), let result: String = numberFormatter.string(for: lPriceInt) else { return }
+        
+        lpriceLabel.text = result + "원"
         
         completionHandler = { [weak self] in
-            self?.repository.deleteLikeItem(savedItemList: savedItem, savedItem: savedItem[index])
+            self?.repository.deleteLikeItem(savedItemList: savedItem,
+                                            savedItem: savedItem[index])
         }
     }
     
@@ -128,7 +141,7 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
         
         likeButton.snp.makeConstraints { make in
             make.bottom.trailing.equalTo(productImage).inset(8)
-            make.size.equalTo(44) //MARK: - 사이즈 조정 가능
+            make.size.equalTo(44)
         }
         
         mallNameLabel.snp.makeConstraints { make in
@@ -157,17 +170,7 @@ class ShoppingListCollectionViewCell: UICollectionViewCell {
         setConstraints()
     }
     
-
-//    override func prepareForReuse() {
-//        super.prepareForReuse()
-//        productImage.image = nil
-//        likeButton.imageView?.image = nil
-//        mallNameLabel.text = nil
-//        titleLabel.text = nil
-//        lpriceLabel.text = nil
-//    }
-    
-    @available(*, unavailable) //MARK: - 이게 여기서 되는게 맞았나...? 아니였던거 같은데...?
+    @available(*, unavailable)
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
